# Reduce allocations while creating a slice of keys from map and fix a …