…typo

- This PR reduces the no. of allocations by ~83% and increases the speed by ~48% which creating the slice while forming the key slice from a map. 
- Also fixed a small typo and added one doc string.

Benchmark script:  https://play.golang.org/p/1Qv2z4W0EMO

Results:
```
➜  dev-vm go test -v -run=^$ -bench=BenchmarkMapSlice a_test.go --count=20 >> {old,new}.txt
➜  dev-vm benchstat old.txt new.txt                                                                  
name        old time/op    new time/op    delta
MapSlice-8    1.29µs ± 4%    0.68µs ± 6%  -47.57%  (p=0.000 n=20+20)

name        old alloc/op   new alloc/op   delta
MapSlice-8    1.01kB ± 0%    0.51kB ± 0%  -49.21%  (p=0.000 n=20+20)

name        old allocs/op  new allocs/op  delta
MapSlice-8      6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=20+20)
```