@sabandi 2020-09-27
Thanks @felixge for the encouragement🙂.  I will keep reading the hotpath code multiple times to see if I can find any optimization there. So far it looks like you didn't leave any room there for optimizations. 🙂