@felixge 2020-09-28
@sabandi yeah, my recent optimizations definitely picked all the low hanging fruits there I think. To make it even faster will probably require hacking on the go runtime itself 🙈.

Anyway, I suppose there are still functional things left to do, e.g. exploring if the raw data could be streamed out, rather than just the aggregation at the end. This might allow integrating with tools such as FlameScope.