@sabandi 2020-09-28
sure @felixge will try to understand more about FlameScope and how it can be integrated.